requirements.txt
biopython==1.65
docopt==0.6.2
numpy==1.9.2
wheel==0.24.0
nose == 1.3.7
rednose == 0.4.1
behave == 1.2.5
scripttest == 1.3
jinja2 == 2.8
pyyaml == 3.11
bcbio-gff
pybedtools
